backend/requirements.txt
python-chess

Flask